Only allow scrolling in details menu
This prevents a sudden shift in wishbanners due to animations causing and then removing the scroll bar. It also removes the ability to scroll on the main screen, which  can be resolved in the future by moving the animations from the top of the screen instead of a set REM above their current position.