Resolve Vue slow update animation flickering
When animationIndex is updated, the src of the image does not immediately change - the classes actually change first, resulting in a brief animation where the image is suddenly darkened and scaled to 200%, which is very distracting. This was resolved by adding a 0.1s delay which is barely noticeable and allows for the image src to update completely.