Fix animation/sfx desync
This commit fully messes up with animation indexes until both the image and audio are ready to be served, removing the old hardcoded animation delay. One issue is that occasionally when repeating an audio piece audio.pause(), setting the current time to zero, and then playing again (particularly near the end of an audio clip) causes a measurable delay and causes another small desync.