Insert ItemObtainScreen after roll
This commit inserts some extra functionality to Inventory's addItemsViaGacha by making it return an admittedly not great format (it really should be the standardised format used in ObtainItem) and passing it along to ObtainScreen via App. This data may be passable to 10pulls but we'll take care of that later likely by rewriting the output from addItemsFromGacha to be more specific.