Add quest screen
This is a major commit that adds a quest screen for dailies/events as well as allows the user to edit just dailies for now. User input is not sanitised nor properly checked for overflow, etc. Rewards cannot yet be modified. Backend linking is complete with a partial rewrite of Inventory to use ItemStringQuantity over string arrays.